requirements.txt
requests==2.32.3
python-dotenv
pytest==8.2.1
jsonschema==4.22.0
allure-pytest==2.13.5
pytest-playwright==0.5.0